pyproject.toml
[project]
name = "mosaic"
version = "0.1.0"
description = "composite objective protein design"
readme = "README.md"
authors = [
    { name = "Nick Boyd", email = "nick@escalante.bio" }
]
requires-python = "==3.12.*"
dependencies = [
    "dm-haiku>=0.0.13",
    "esm2quinox==0.1.0",
    "flax>=0.10.2",
    "gemmi>=0.6.5",
    "ipymolstar>=0.0.9",
    "jax>=0.8.1",
    "joltz",
    "marimo>=0.10.9",
    "matplotlib>=3.10.0",
    "ml-collections>=1.0.0",
    "optax>=0.2.4",
    "esmj",
    "scipy==1.15.1",
    "ablang>=0.3.1",
    "jablang",
    "torchvision>=v0.22.0",
    "equinox>=0.13.0",
    "torch==2.7.1",
    "numpy==2.1",
    "protenix",
    "httpx>=0.28.1",
    "joltzgen",
    "nvidia-cublas-cu12>=12.9.1.4",
    "nvidia-cusolver-cu12==11.7.3.90",
    "molviewspec>=1.8.1",
    "wandb>=0.18.7",
    "python-dotenv>=1.2.1",
    "plotly>=6.5.2",
]

[dependency-groups]
dev = [
    "ty>=0.0.14",
]
jax-cpu = [
  "jax[cpu]",
]
jax-cuda = [
  "jax[cuda]",
]
jax-tpu= [
  "jax[tpu]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
artifacts = [
    "src/mosaic/proteinmpnn/weights/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mosaic", "vhh"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
joltz = { git = "https://github.com/nboyd/joltz" }
esmj = { git = "https://github.com/escalante-bio/esmj" }
protenix = { git = "https://github.com/escalante-bio/protenij" }
jablang = { git = "https://github.com/escalante-bio/jablang" }
joltzgen = { git = "https://github.com/escalante-bio/joltzgen.git" }

[tool.uv]
conflicts = [
    [
      { group = "jax-cpu" },
      { group = "jax-cuda" },
      { group = "jax-tpu" },
    ],
]
package = true
override-dependencies = ["scipy==1.15.1", "numpy==2.1"]
